msm: kgsl: Fix gpuaddr_in_range() to check upper bound
Currently gpuaddr_in_range() accepts only the gpuaddr & returns
true if it lies in valid range. But this does not mean that the
entire buffer is within range.
Modify the function to accept size as a parameter and check that
both starting & ending points of buffer lie within mmu range.

Change-Id: I1d722295b9a27e746bfdb6d3bf409ffe722193cb
Author: Commit c9b3b88b767cca85fe504289edf31eb1599c21f4 authored on 2021-12-07 07:20:00 by Abhishek Barman <quic_abarman@quicinc.com>. Committed by QU TAO <quic_ksanadhy@quicinc.com> on 2022-05-26 13:16:00